Edit file relocation
